requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
python-whois==0.7.3
dnspython==2.1.0
shodan==1.25.0
